fix: API key scope 默认值改为 shared，API 路径重命名为 /api/api-keys
- Migration 047 scope 默认值从 'remote' 改为 'shared'，并修复已有数据
- Schema 文件补充缺失的 scope/priority/weight 列
- API key 路由从 remote_bp 移至独立 api_keys_bp（/api/api-keys）
- 后端 store_api_key fallback 默认值改为 'shared'
- 前端 API 路径和 React Query key 同步更新
- 测试文件中的 API 路径和前端页面路径同步修复
