+++ schema/schema-postgres.sql
@@ -136,7 +136,10 @@ CREATE TABLE api_key_store (
     created_at timestamp without time zone DEFAULT CURRENT_TIMESTAMP,
     updated_at timestamp without time zone DEFAULT CURRENT_TIMESTAMP,
     cli_tools text,
-    cli_settings text
+    cli_settings text,
+    scope text DEFAULT 'shared',
+    priority integer DEFAULT 0,
+    weight integer DEFAULT 100
 );
 
 CREATE SEQUENCE api_key_store_id_seq
